+++ interpreter/src/main/org/trypticon/hex/interpreters/primitives/UInt.java
@@ -56,7 +56,7 @@ public class UInt extends Number implements Value {
         return value;
     }
 
-    public int length() {
+    public long length() {
         return 4;
     }
